Be clearer and more consistent about args.artifacts_root
